More int to NSInteger updates
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40468 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSThemeDrawing.m
@@ -76,9 +76,9 @@
 
 @interface NSTableView (Private)
 - (CGFloat *)_columnOrigins;
-- (void) _willDisplayCell: (NSCell*)cell forTableColumn: (NSTableColumn *)tb row: (int)index;
+- (void) _willDisplayCell: (NSCell*)cell forTableColumn: (NSTableColumn *)tb row: (NSInteger)index;
 // TESTPLANT-MAL-2016: Keeping for tableview grouped row support AKA Cocoa...
-- (NSCell *) _dataCellForTableColumn: (NSTableColumn *)tb row: (int) rowIndex;
+- (NSCell *) _dataCellForTableColumn: (NSTableColumn *)tb row: (NSInteger) rowIndex;
 - (BOOL)_isGroupRow: (NSInteger)rowIndex;
 @end
